feat: connect SDK
